+++ unified-codes/libs/data/README.md
@@ -1,0 +1,7 @@
+# data
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `ng test data` to execute the unit tests via [Jest](https://jestjs.io).
