+++ unified-codes/libs/data/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  name: 'util',
+  name: 'data',
   preset: '../../jest.config.js',
   globals: {
     'ts-jest': {
@@ -10,5 +10,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-  coverageDirectory: '../../coverage/libs/util',
+  coverageDirectory: '../../coverage/libs/data',
 };
