+++ unified-codes/libs/data/src/lib/types/Entity.ts
