+++ unified-codes/libs/data/src/lib/types/EntityNode.ts
