+++ unified-codes/libs/data/src/lib/types/Property.ts
