+++ unified-codes/libs/ui/src/lib/entity/molecules/EntityTableRow.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { EntityTableRow } from './EntityTableRow';
 
-import { Entity, EntityNode } from '@unified-codes/util';
+import { Entity, EntityNode } from '@unified-codes/data';
 
 export default {
   component: EntityTableRow,
