--- unified-codes/libs/util/README.md
@@ -1,7 +1,0 @@
-# util
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `ng test util` to execute the unit tests via [Jest](https://jestjs.io).
