+++ unified-codes/nx.json
@@ -38,7 +38,7 @@
     "ui": {
       "tags": []
     },
-    "util": {
+    "data": {
       "tags": []
     }
   }
