+++ unified-codes/workspace.json
@@ -295,9 +295,9 @@
         }
       }
     },
-    "util": {
-      "root": "libs/util",
-      "sourceRoot": "libs/util/src",
+    "data": {
+      "root": "libs/data",
+      "sourceRoot": "libs/data/src",
       "projectType": "library",
       "schematics": {},
       "architect": {
@@ -305,15 +305,15 @@
           "builder": "@nrwl/linter:lint",
           "options": {
             "linter": "eslint",
-            "tsConfig": ["libs/util/tsconfig.lib.json", "libs/util/tsconfig.spec.json"],
-            "exclude": ["**/node_modules/**", "!libs/util/**/*"]
+            "tsConfig": ["libs/data/tsconfig.lib.json", "libs/data/tsconfig.spec.json"],
+            "exclude": ["**/node_modules/**", "!libs/data/**/*"]
           }
         },
         "test": {
           "builder": "@nrwl/jest:jest",
           "options": {
-            "jestConfig": "libs/util/jest.config.js",
-            "tsConfig": "libs/util/tsconfig.spec.json",
+            "jestConfig": "libs/data/jest.config.js",
+            "tsConfig": "libs/data/tsconfig.spec.json",
             "passWithNoTests": true
           }
         }
